refactor: separate business logic from listener to service class
